requirements.txt
confluent-kafka
snowflake-connector-python
streamlit
pandas
hypothesis
pytest
